Make max number of simultaneous processes configurable
That limit has always existed but was implicit and undocumented. It is
caused by the fact that every running process will block one worker
thread from the ThreadPoolExecutor (passed to run_in_executor()).
As we used the default executor, that limit could vary between python
versions (32 on py37 but only 8 on py38).

We now use our own ThreadPoolExecutor instead of the loop's default one.
That way we can make the max number of processes configurable and avoid
interfering with other code that uses the default executor.